requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
eli5==0.11.0
feedparser==6.0.8
graphviz==0.16
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
Jinja2==2.11.2
joblib==1.1.0
MarkupSafe==1.1.1
memory-profiler==0.60.0
numpy==1.20.2
packaging==21.3
pandas==1.1.3
pluggy==1.0.0
psutil==5.8.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0
reader==2.6
requests==2.26.0
scikit-learn==0.23.2
scipy==1.5.2
scripts==2.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.3.1
stua==0.2
tabulate==0.8.9
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
